requirements.txt
requests==2.32.3
sentence-transformers==4.1.0
pinecone==6.0.2
google-generativeai==0.8.5
python-dotenv==1.1.0